backend/requirements.txt
Flask==3.1.2
Flask-Cors==6.0.2
PyJWT==2.10.1
requests==2.32.5
uuid==1.30
Authlib==1.6.6
python3-saml==1.16.0